fix(system): 修正 SMC 结构体 ABI 错位、CPU 采样溢出与菜单栏指标重复
- SMCParamStruct.padding 由 UInt16 恢复为 UInt8：UInt16 使 result/status/
  data8 偏移整体后移 2 字节，驱动读不到命令码，全部 SMC 读取静默失败；
  新增 SMCParamStructLayoutTests 锁定 stride=80 与关键偏移契约
- CPUUsageSampler 的 tick 求和/差分统一升位 UInt64：UInt32 域在开机
  累计数十天后必溢出 trap；补近上限回归测试
- 菜单栏独立状态项模式下主图标不再重复拼接全部指标（metrics 由独立
  item 承载后主 item 只留 countdown）；顺带合并 composeMainTitle
  重复分支

+++ Sources/OmniForgeSMC/SMCClient.swift
@@ -119,12 +119,16 @@ public final class SMCClient: SMCReading {
     }
 }
 
+/// 与 AppleSMC 驱动用户态 ABI 逐字段对齐的经典布局。
+/// `padding` 必须是单个 `UInt8`：若放宽为 `UInt16`，`result`/`status`/`data8`
+/// 的偏移会整体后移 2 字节（驱动在 data8@40 读命令码、在 result@38 写返回码），
+/// 所有 SMC 调用静默失败。布局契约由 SMCParamStructLayoutTests 锁定。
 struct SMCParamStruct {
     var key: UInt32 = 0
     var vers = SMCVersion()
     var pLimitData = SMCPLimitData()
     var keyInfo = SMCKeyInfoData()
-    var padding: UInt16 = 0
+    var padding: UInt8 = 0
     var result: UInt8 = 0
     var status: UInt8 = 0
     var data8: UInt8 = 0
